# apispec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/training/yolo/result": {
      "post": {
        "tags": [
          "Training",
          "Training"
        ],
        "summary": "Yolo Training Result",
        "description": "Handle the result of YOLO training.",
        "operationId": "yolo_training_result_training_yolo_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoloHandlingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/validation/yolo/validate/detection": {
      "post": {
        "tags": [
          "Validation",
          "Validation"
        ],
        "summary": "Start Validation",
        "operationId": "start_validation_validation_yolo_validate_detection_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoloDetValidationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/validation/yolo/{vid}": {
      "get": {
        "tags": [
          "Validation",
          "Validation"
        ],
        "summary": "Get Validation Info",
        "description": "Get validation status and results",
        "operationId": "get_validation_info_validation_yolo__vid__get",
        "parameters": [
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/validation/yolo/result": {
      "post": {
        "tags": [
          "Validation",
          "Validation"
        ],
        "summary": "Receive Result",
        "description": "Receive validation completion callback and persist to Mongo/MinIO.",
        "operationId": "receive_result_validation_yolo_result_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Result"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "YoloDetValidationParams": {
        "properties": {
          "imgsz": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imgsz",
            "default": 640
          },
          "batch": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch",
            "default": 32
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device",
            "default": "cpu"
          },
          "workers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workers",
            "default": 8
          },
          "conf": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conf",
            "default": 0.001
          },
          "iou": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iou",
            "default": 0.6
          },
          "max_det": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Det",
            "default": 300
          },
          "save_json": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Save Json",
            "default": true
          },
          "save_txt": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Save Txt",
            "default": true
          },
          "save_conf": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Save Conf",
            "default": true
          },
          "plots": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plots",
            "default": true
          },
          "verbose": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verbose",
            "default": true
          },
          "half": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Half",
            "default": false
          },
          "dnn": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dnn",
            "default": false
          },
          "agnostic_nms": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agnostic Nms",
            "default": false
          },
          "augment": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Augment",
            "default": false
          },
          "rect": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rect",
            "default": false
          }
        },
        "type": "object",
        "title": "YoloDetValidationParams",
        "description": "model은 best.pt로 고정"
      },
      "YoloDetValidationRequest": {
        "properties": {
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "tid": {
            "type": "string",
            "title": "Tid"
          },
          "cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cid"
          },
          "did": {
            "type": "string",
            "title": "Did"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "description": "Task type for YOLO validation (detection/segmentation/pose/obb/classification)"
          },
          "parameters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/YoloDetValidationParams"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "pid",
          "tid",
          "did",
          "task_type",
          "parameters"
        ],
        "title": "YoloDetValidationRequest"
      },
      "YoloHandlingRequest": {
        "properties": {
          "workdir": {
            "type": "string",
            "title": "Workdir"
          },
          "result": {
            "$ref": "#/components/schemas/YoloTrainingResult"
          }
        },
        "type": "object",
        "required": [
          "workdir",
          "result"
        ],
        "title": "YoloHandlingRequest"
      },
      "YoloTrainingResult": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "origin_tid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Tid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "workdir": {
            "type": "string",
            "title": "Workdir"
          },
          "started_time": {
            "type": "string",
            "title": "Started Time"
          },
          "completed_time": {
            "type": "string",
            "title": "Completed Time"
          },
          "parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          },
          "did": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Did"
          },
          "cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cid"
          },
          "artifacts_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts Path"
          },
          "error_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Details"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "pid",
          "status",
          "task_type",
          "workdir",
          "started_time",
          "completed_time"
        ],
        "title": "YoloTrainingResult"
      }
    }
  }
}